Fix linkedin icon size on Firefox
Linkedin icon on Firefox is much smaller than the other icon, and does
not fit its enclosing circle. The likely reason for this is that Firefox
is respecting the "width" and "height" properties on the symbol element.

This commit removes them so the symbol takes the whole circle, as the
other icons.